made module structure more clear
removed `transformers`: everything is a transformer
moved everything from `transformers` to `preprocessing`
moved selectors in preprocessing to `selectors`
moved EstimatorTransformer to `meta`
