+++ sklego/preprocessing.py
@@ -1,148 +1,62 @@
-import pandas as pd
+import numpy as np
+from patsy import dmatrix, build_design_matrices, PatsyError
 from sklearn.base import BaseEstimator, TransformerMixin
-from sklearn.utils.validation import check_is_fitted
+from sklearn.utils import check_array, check_X_y
+from sklearn.utils.validation import FLOAT_DTYPES, check_random_state, check_is_fitted
+
+from sklego.common import TrainOnlyTransformerMixin
 
 
-class PandasTypeSelector(BaseEstimator, TransformerMixin):
+class RandomAdder(TrainOnlyTransformerMixin, BaseEstimator):
+    def __init__(self, noise=1, random_state=None):
+        self.noise = noise
+        self.random_state = random_state
+
+    def fit(self, X, y):
+        super().fit(X, y)
+        X, y = check_X_y(X, y, estimator=self, dtype=FLOAT_DTYPES)
+        self.dim_ = X.shape[1]
+
+        return self
+
+    def transform_train(self, X):
+        rs = check_random_state(self.random_state)
+        check_is_fitted(self, ['dim_'])
+
+        X = check_array(X, estimator=self, dtype=FLOAT_DTYPES)
+
+        return X + rs.normal(0, self.noise, size=X.shape)
+
+
+class PatsyTransformer(TransformerMixin, BaseEstimator):
     """
-    Select columns in a pandas dataframe based on their dtype
+    The patsy transformer offers a method to select the right columns
+    from a dataframe as well as a DSL for transformations. It is inspired
+    from R formulas.
 
-    :param include: types to be included in the dataframe
-    :param exclude: types to be exluded in the dataframe
+    This is can be useful as a first step in the pipeline.
+
+    :param formula: a patsy-compatible formula
     """
-    def __init__(self, include=None, exclude=None):
-        self.include = include
-        self.exclude = exclude
+
+    def __init__(self, formula):
+        self.formula = formula
 
     def fit(self, X, y=None):
-        """
-        Saves the column names for check during transform
-
-        :param X: pandas dataframe to select dtypes out of
-        :param y: not used in this class
-        """
-        self._check_X_for_type(X)
-        self.type_columns_ = list(X.select_dtypes(include=self.include, exclude=self.exclude))
-
-        if len(self.type_columns_) == 0:
-            raise ValueError(f'Provided type(s) results in empty dateframe')
-
+        """Fits the estimator"""
+        X_ = dmatrix(self.formula, X)
+        assert np.array(X_).shape[0] == np.array(X).shape[0]
+        self.design_info_ = X_.design_info
         return self
 
     def transform(self, X):
         """
-        Transforms pandas dataframe by (de)selecting columns based on their dtype
+        Applies the formula to the matrix/dataframe X.
 
-        :param X: pandas dataframe to select dtypes for
+        Returns an design array that can be used in sklearn pipelines.
         """
-        check_is_fitted(self, 'type_columns_')
-
-        self._check_X_for_type(X)
-
-        transformed_df = X.select_dtypes(include=self.include, exclude=self.exclude)
-
-        if set(list(transformed_df)) != set(self.type_columns_):
-            raise ValueError(f'Columns were not equal during fit and transform')
-
-        return transformed_df
-
-    @staticmethod
-    def _check_X_for_type(X):
-        """Checks if input of the Selector is of the required dtype"""
-        if not isinstance(X, pd.DataFrame):
-            raise TypeError("Provided variable X is not of type pandas.DataFrame")
-
-
-class ColumnSelector(BaseEstimator, TransformerMixin):
-    """
-    Allows selecting specific columns from a pandas DataFrame by name. Can be useful in a sklearn Pipeline.
-
-    :param columns: column name ``str`` or list of column names to be selected
-
-    .. note::
-        Raises a ``TypeError`` if input provided is not a DataFrame
-
-        Raises a ``ValueError`` if columns provided are not in the input DataFrame
-
-    :Example:
-
-    >>> # Selecting a single column from a pandas DataFrame
-    >>> import pandas as pd
-    >>> df = pd.DataFrame({
-    ...     'name': ['Swen', 'Victor', 'Alex'],
-    ...     'length': [1.82, 1.85, 1.80],
-    ...     'shoesize': [42, 44, 45]
-    ... })
-    >>> ColumnSelector(['length']).fit_transform(df)
-       length
-    0    1.82
-    1    1.85
-    2    1.80
-
-    >>> # Selecting multiple columns from a pandas DataFrame
-    >>> ColumnSelector(['length', 'shoesize']).fit_transform(df)
-       length  shoesize
-    0    1.82        42
-    1    1.85        44
-    2    1.80        45
-
-    >>> # Selecting non-existent columns returns in a KeyError
-    >>> ColumnSelector(['weight']).fit_transform(df)
-    Traceback (most recent call last):
-        ...
-    KeyError: "['weight'] column(s) not in DataFrame"
-
-    >>> # How to use the ColumnSelector in a sklearn Pipeline
-    >>> from sklearn.pipeline import Pipeline
-    >>> from sklearn.preprocessing import StandardScaler
-    >>> pipe = Pipeline([
-    ...     ('select', ColumnSelector(['length'])),
-    ...     ('scale', StandardScaler()),
-    ... ])
-    >>> pipe.fit_transform(df)
-    array([[-0.16222142],
-           [ 1.29777137],
-           [-1.13554995]])
-    """
-
-    def __init__(self, columns: list):
-        # if the columns parameter is not a list, make it into a list
-        if not isinstance(columns, list):
-            columns = [columns]
-
-        self.columns = columns
-
-    def fit(self, X, y=None):
-        """
-        Checks 1) if input is a DataFrame, and 2) if column names are in this DataFrame
-
-        :param X: ``pd.DataFrame`` on which we apply the column selection
-        :param y: ``pd.Series`` labels for X. unused for column selection
-        :returns: ``ColumnSelector`` object.
-        """
-
-        self._check_X_for_type(X)
-        self._check_column_names(X)
-        return self
-
-    def transform(self, X):
-        """Returns a pandas DataFrame with only the specified columns
-
-        :param X: ``pd.DataFrame`` on which we apply the column selection
-        :returns: ``pd.DataFrame`` with only the selected columns
-        """
-        if self.columns:
-            return X[self.columns]
-        return X
-
-    def _check_column_names(self, X):
-        """Check if one or more of the columns provided doesn't exist in the input DataFrame"""
-        non_existent_columns = set(self.columns).difference(X.columns)
-        if len(non_existent_columns) > 0:
-            raise KeyError(f'{list(non_existent_columns)} column(s) not in DataFrame')
-
-    @staticmethod
-    def _check_X_for_type(X):
-        """Checks if input of the Selector is of the required dtype"""
-        if not isinstance(X, pd.DataFrame):
-            raise TypeError("Provided variable X is not of type pandas.DataFrame")
+        check_is_fitted(self, 'design_info_')
+        try:
+            return build_design_matrices([self.design_info_], X)[0]
+        except PatsyError as e:
+            raise RuntimeError from e
