+++ tests/test_selectors/test_pandastypeselector.py
@@ -2,7 +2,7 @@ import pytest
 import pandas as pd
 import itertools as it
 import numpy as np
-from sklego.preprocessing import PandasTypeSelector
+from sklego.selectors import PandasTypeSelector
 from tests.conftest import id_func
